requirements.txt
tkinter
streamlit
pycryptodome
pytest